requirements/base.txt
asgiref==3.7.2
Django==5.0.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-jazzmin==2.6.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
